theme/react/test/backend/requirements.txt
Flask
flask-cors
google-generativeai
Pillow
grpcio
